@@ -2,8 +2,7 @@
   "name": "example-webpack",
   "private": true,
   "scripts": {
-    "dev": "webpack-dev-server examples/simple/index.ts",
-    "dev:react": "webpack-dev-server examples/react/index.ts"
+    "dev": "webpack-dev-server ./src/index.ts"
   },
   "author": "",
   "license": "ISC",
@@ -15,14 +14,16 @@
     "@babel/core": "^7.19.6",
     "@babel/preset-env": "^7.19.4",
     "@babel/preset-typescript": "^7.18.6",
+    "assert": "^2.1.0",
     "babel-loader": "^8.2.3",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.2.0",
     "dotenv-webpack": "^8.0.1",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.2",
+    "path-browserify": "^1.0.1",
     "style-loader": "^3.2.1",
-    "typescript": "^4.4.4",
+    "typescript": "^5",
     "webpack": "^5.64.1",
     "webpack-cli": "^4.9.1",
     "webpack-dev-server": "^4.3.1"
